+++ web/assets/js/LiteMol-core.js
@@ -11057,7 +11057,7 @@ var LiteMol;
 (function (LiteMol) {
     var Core;
     (function (Core) {
-        Core.VERSION = { number: "2.4.9", date: "Nov 12 2016" };
+        Core.VERSION = { number: "2.4.10", date: "Nov 12 2016" };
     })(Core = LiteMol.Core || (LiteMol.Core = {}));
 })(LiteMol || (LiteMol = {}));
 /*
@@ -12128,8 +12128,13 @@ var LiteMol;
                     function isResidueNucleotide(atoms, residues, entities, index) {
                         if (aminoAcidNames[residues.name[index]] || entities.entityType[residues.entityIndex[index]] !== Core.Structure.EntityType.Polymer)
                             return false;
-                        var o5 = false, c3 = false, n3 = false, names = atoms.name, assigned = 0;
-                        for (var i = residues.atomStartIndex[index], max = residues.atomEndIndex[index]; i < max; i++) {
+                        var o5 = false, c3 = false, n3 = false, p = false, names = atoms.name, assigned = 0;
+                        var start = residues.atomStartIndex[index], end = residues.atomEndIndex[index];
+                        // test for single atom instances
+                        if (end - start === 1 && !residues.isHet[start] && names[start] === 'P') {
+                            return true;
+                        }
+                        for (var i = start; i < end; i++) {
                             var n = names[i];
                             if (!o5 && n === "O5'") {
                                 o5 = true;
@@ -12143,10 +12148,14 @@ var LiteMol;
                                 n3 = true;
                                 assigned++;
                             }
-                            if (assigned === 3)
+                            else if (!p && n === 'P') {
+                                p = true;
+                                assigned++;
+                            }
+                            if (assigned === 4)
                                 break;
                         }
-                        return o5 && c3 && n3;
+                        return o5 && c3 && n3 && p;
                     }
                     function analyzeSecondaryStructure(atoms, residues, entities, start, end, elements) {
                         var asymId = residues.asymId, entityIndex = residues.entityIndex, currentType = 0 /* None */, currentElementStartIndex = start, currentResidueIndex = start, residueCount = end;